SimVP 2 Classification. Preprocess updated
Preprocess script organized. Now test image will be processed in a different script, during inference, because it needs to be extracted in order.
Now images are being grouped by trimester tasking the max what I wasn't doing before.

SimVP model in OpenSTL is being modified to be a classificator.
Needed to change the last layer inside SimVP class. Also needed to change the loss and adjust metrics for classification
instead of regression. Still in progress.

Still need to do the test script